straighten serialization and deserialization for transaction
